Performance Improvements
Re-written. Now unless the last time it checked was 0 days ago, it will check and update a file with the last ran date and the result.

This should mean for faster execution and fewer issues with resource load.

Probably a better more elegant way to do this, but this works.