Update yum-outstanding-security-updates-check.sh
Found a number of issues with the temp file again that the checks weren't catching.

For example it wouldn't work if the file was:

1970-01-01 00:0
0
0

Due to the original count of the lines checking it was less than 2. As the above is three, it wasn't overwriting it.

In addition. I found if the first item isn't a date, maybe:

0
0

then it would return 0 anyway, regardless. I've fixed this in a round about way. There are performance gains to be had in this and less breaky ways of doing it, so feel free to suggest.