Variable fix
Bug fixed where if you were to manually delete the temp file and re-run the script, for example:

rm -f /tmp/lastyumcheck; /var/prtg/scripts/yum-outstanding-security-updates-check.sh

It would output this:

0:9999:9999 Outstanding Security Updates

Which might look a little scary. This was due to the OUTSTANDING variable not being set again in the list if then else section. Added one line.